requirements.txt
boto3==1.26.156
botocore==1.29.156
jmespath==1.0.1
psycopg2==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.1
six==1.16.0
tqdm==4.65.0
urllib3==1.26.16
